fixup! [c++] Protocol Buffers writer
